✅[CHORE] UI Component 분리
