+++ Assignment/Youtube-iOS/Youtube-iOS/Sources/VCs/Youtube/VCs/SubscriptionsVC.swift
@@ -11,42 +11,14 @@ class SubscriptionsVC: UIViewController {
     
     // MARK: - UI
     
-    private var topView = UIView()
-    
-    private var logoImageView = UIImageView()
-    
-    private var buttonStackView = UIStackView()
-    private var windowSharingButton = UIButton()
-    private var notificationButton = UIButton()
-    private var searchButton = UIButton()
-    private var profileButton = UIButton()
-    
-    private var channelCollectionView: UICollectionView = {
-        let layout = UICollectionViewFlowLayout()
-        layout.scrollDirection = .horizontal
-        let collectionView = UICollectionView(frame: .zero, collectionViewLayout: layout)
-        collectionView.isScrollEnabled = true
-        collectionView.showsHorizontalScrollIndicator = false
-        return collectionView
-    }()
-    
-    private var sortCollectionView: UICollectionView = {
-        let layout = UICollectionViewFlowLayout()
-        layout.scrollDirection = .horizontal
-        let collectionView = UICollectionView(frame: .zero, collectionViewLayout: layout)
-        collectionView.isScrollEnabled = true
-        collectionView.showsHorizontalScrollIndicator = false
-        return collectionView
-    }()
+    private var topView = TopView()
+    private var collectionViews = CollectionViews()
     
     private var videoTableView = UITableView()
 
     // MARK: - Properties
     
     private var videoList = [VideoDataModel]()
-    private var channelList = [ChannelDataModel]()
-    
-    private var sortList = [String]()
 
     // MARK: - Life Cycle
     
@@ -58,7 +30,6 @@ class SubscriptionsVC: UIViewController {
         
         setData()
         setTableView()
-        setCollectionView()
     }
 }
 
@@ -67,61 +38,24 @@ class SubscriptionsVC: UIViewController {
 extension SubscriptionsVC {
     private func initUI() {
         view.backgroundColor = .white
-        
-        logoImageView.image = UIImage(named: "logo")
-        logoImageView.contentMode = .scaleAspectFill
-        
-        buttonStackView.alignment = .fill
-        buttonStackView.axis = .horizontal
-        buttonStackView.distribution = .fillEqually
-        buttonStackView.spacing = 5
-        buttonStackView.addArrangedSubview(windowSharingButton)
-        buttonStackView.addArrangedSubview(notificationButton)
-        buttonStackView.addArrangedSubview(searchButton)
-        buttonStackView.addArrangedSubview(profileButton)
-        
-        windowSharingButton.setImage(UIImage(named: "windowSharingIcon"), for: .normal)
-        notificationButton.setImage(UIImage(named: "notificationIcon"), for: .normal)
-        searchButton.setImage(UIImage(named: "searchIcon"), for: .normal)
-        profileButton.setImage(UIImage(named: "wesoptProfile"), for: .normal)
     }
     
     private func setConstraints() {
-        view.addSubviews([topView, videoTableView])
-        topView.addSubviews([logoImageView, buttonStackView, channelCollectionView, sortCollectionView])
+        view.addSubviews([topView, collectionViews, videoTableView])
         
         topView.snp.makeConstraints { make in
             make.top.leading.trailing.equalTo(view.safeAreaLayoutGuide)
-            make.height.equalTo(184)
+            make.height.equalTo(44)
         }
         
-        logoImageView.snp.makeConstraints { make in
-            make.top.equalToSuperview()
-            make.leading.equalToSuperview().inset(16)
-            make.width.equalTo(96)
-            make.height.equalTo(20)
-        }
-        
-        buttonStackView.snp.makeConstraints { make in
-            make.trailing.equalToSuperview().inset(16)
-            make.centerY.equalTo(logoImageView.snp.centerY)
-            make.width.equalTo(160)
-        }
-        
-        channelCollectionView.snp.makeConstraints { make in
-            make.top.equalTo(buttonStackView.snp.bottom)
+        collectionViews.snp.makeConstraints { make in
+            make.top.equalTo(topView.snp.bottom)
             make.leading.trailing.equalToSuperview()
-            make.height.equalTo(104)
-        }
-
-        sortCollectionView.snp.makeConstraints { make in
-            make.top.equalTo(channelCollectionView.snp.bottom)
-            make.leading.trailing.equalToSuperview()
-            make.height.equalTo(40)
+            make.height.equalTo(154)
         }
         
         videoTableView.snp.makeConstraints { make in
-            make.top.equalTo(topView.snp.bottom)
+            make.top.equalTo(collectionViews.snp.bottom)
             make.leading.trailing.bottom.equalTo(view.safeAreaLayoutGuide)
         }
     }
@@ -135,21 +69,6 @@ extension SubscriptionsVC {
             VideoDataModel(thumbnailImage: "wesoptiOSPart", profileImage: "wesoptProfile", title: "5주차 세미나 어쩌고 저쩌고 재밌다 아요 지대 멋지다 지대 짱", channelName: "WE SOPT", views: 100, date: 3),
             VideoDataModel(thumbnailImage: "wesoptiOSPart", profileImage: "wesoptProfile", title: "6주차 세미나 어쩌고 저쩌고 재밌다 아요 지대 멋지다 지대 짱", channelName: "WE SOPT", views: 100, date: 3)
         ])
-        
-        channelList.append(contentsOf: [
-            ChannelDataModel(channelImage: "ggamju1", channelName: "배고파"),
-            ChannelDataModel(channelImage: "ggamju2", channelName: "졸려"),
-            ChannelDataModel(channelImage: "ggamju3", channelName: "아직도"),
-            ChannelDataModel(channelImage: "ggamju4", channelName: "시험이"),
-            ChannelDataModel(channelImage: "ggamju5", channelName: "안끝난"),
-            ChannelDataModel(channelImage: "ggamju6", channelName: "사람이"),
-            ChannelDataModel(channelImage: "ggamju7", channelName: "있다?"),
-            ChannelDataModel(channelImage: "ggamju8", channelName: "바로 나")
-        ])
-        
-        sortList.append(contentsOf: [
-            "전체", "오늘", "이어서 시청하기", "시청하지 않음", "실시간", "게시물"
-        ])
     }
     
     private func setTableView() {
@@ -160,23 +79,6 @@ extension SubscriptionsVC {
         videoTableView.register(videoNib, forCellReuseIdentifier: VideoTVC.identifier)
         
         videoTableView.allowsSelection = false
-    }
-    
-    private func setCollectionView() {
-        channelCollectionView.delegate = self
-        channelCollectionView.dataSource = self
-        channelCollectionView.register(ChannelCVC.self, forCellWithReuseIdentifier: ChannelCVC.identifier)
-        
-        sortCollectionView.delegate = self
-        sortCollectionView.dataSource = self
-        sortCollectionView.register(SortCVC.self, forCellWithReuseIdentifier: SortCVC.identifier)
-    }
-    
-    private func setLabelWidth(index: Int) -> CGFloat {
-        let label = UILabel()
-        label.text = sortList[index]
-        label.sizeToFit()
-        return label.frame.width + 10 + 10
     }
 }
 
@@ -201,57 +103,5 @@ extension SubscriptionsVC: UITableViewDataSource {
         let data = videoList[indexPath.row]
         cell.setData(thumnailImage: data.thumbnailImage, profileImage: data.profileImage, title: data.title, channelName: data.channelName, views: data.views, date: data.date)
         return cell
-    }
-}
-
-// MARK: - UICollectionView Delegate
-
-extension SubscriptionsVC: UICollectionViewDelegateFlowLayout {
-    func collectionView(_ collectionView: UICollectionView, layout collectionViewLayout: UICollectionViewLayout, sizeForItemAt indexPath: IndexPath) -> CGSize {
-        if collectionView == channelCollectionView {
-            return CGSize(width: 72, height: 104)
-        } else {
-            return CGSize(width: setLabelWidth(index: indexPath.item), height: 30)
-        }
-    }
-    
-    func collectionView(_ collectionView: UICollectionView, layout collectionViewLayout: UICollectionViewLayout, minimumLineSpacingForSectionAt section: Int) -> CGFloat {
-        return 0
-    }
-    
-    func collectionView(_ collectionView: UICollectionView, layout collectionViewLayout: UICollectionViewLayout, minimumInteritemSpacingForSectionAt section: Int) -> CGFloat {
-        if  collectionView == sortCollectionView {
-            return 9
-        }
-        return 0
-    }
-    
-    func collectionView(_ collectionView: UICollectionView, layout collectionViewLayout: UICollectionViewLayout, insetForSectionAt section: Int) -> UIEdgeInsets {
-        if  collectionView == sortCollectionView {
-            return UIEdgeInsets(top: 0, left: 10, bottom: 0, right: 10)
-        }
-        return UIEdgeInsets.zero
-    }
-}
-
-extension SubscriptionsVC:  UICollectionViewDataSource {
-    func collectionView(_ collectionView: UICollectionView, numberOfItemsInSection section: Int) -> Int {
-        if collectionView == channelCollectionView {
-            return channelList.count
-        } else {
-            return sortList.count
-        }
-    }
-    
-    func collectionView(_ collectionView: UICollectionView, cellForItemAt indexPath: IndexPath) -> UICollectionViewCell {
-        if collectionView == channelCollectionView {
-            guard let cell = collectionView.dequeueReusableCell(withReuseIdentifier: ChannelCVC.identifier, for: indexPath) as? ChannelCVC else { return UICollectionViewCell() }
-            cell.setData(channelImage: channelList[indexPath.row].channelImage, channelName: channelList[indexPath.row].channelName)
-            return cell
-        } else {
-            guard let cell = collectionView.dequeueReusableCell(withReuseIdentifier: SortCVC.identifier, for: indexPath) as? SortCVC else { return UICollectionViewCell() }
-            cell.setLabel(sort: sortList[indexPath.row])
-            return cell
-        }
     }
 }
